Move configuration under common.
