+++ common/configuration/test/basic-config1.json
